equihash: Move `Params` into an internal submodule
